Time: 2444 ms (5.03%), Space: 7.8 MB (54.39%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPerfectSquare(int num) {
-        for(long i = 0 ; i <=num;i++){
+        for(long long  i = 0 ; i <=num;i++){
             long long ans = i * i;
             if(ans == num){
                 return true;
